add public BlockContext.current and BlockContext.push
These allow an ExecutionContext or other code to override
the BlockContext for the current thread. With this
API, the BlockContext is customizable without
creating a new pool of threads.

BlockContext.current is needed to obtain the previous
BlockContext before you push, so you can "chain up" to
it if desired.

BlockContext.push is used to override the context
for a given piece of code.